requirements.txt
cltk-readers @ git+https://github.com/diyclassics/cltk_readers.git@f1d22b4863f87895dbfc58167e1fd36f46d376b5
lime==0.2.0.1
matplotlib==3.7.2
natsort==8.2.0
nltk==3.8.1
pandas==2.1.1
scikit-learn==1.2.0
scipy==1.11.2
seaborn==0.13.0
spacy==3.6.1
tabulate==0.9.0
textacy==0.13.0
tqdm==4.66.1
wordcloud==1.9.2
